Rename enhancements.stretching to enhancements.contrast
