+++ satpy/enhancements/contrast.py
