+++ satpy/etc/enhancements/ahi.yaml
@@ -4,7 +4,7 @@ enhancements:
     standard_name: airmass
     operations:
       - name: stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs:
           stretch: crude
           min_stretch: [-26.2, -43.2, 243.9]
@@ -15,13 +15,13 @@ enhancements:
     sensor: ahi
     operations:
       - name: stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs:
           stretch: crude
           min_stretch: [-36, -1, -80]
           max_stretch: [5, 61, 26]
       - name: gamma
-        method: !!python/name:satpy.enhancements.stretching.gamma
+        method: !!python/name:satpy.enhancements.contrast.gamma
         kwargs:
           gamma: [1.0, 0.5, 0.95]
 
@@ -30,12 +30,12 @@ enhancements:
     sensor: ahi
     operations:
       - name: stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs:
           stretch: crude
           min_stretch: [-7.1, -2.9, 273]
           max_stretch: [ 2.4,  1.1, 300]
       - name: gamma
-        method: !!python/name:satpy.enhancements.stretching.gamma
+        method: !!python/name:satpy.enhancements.contrast.gamma
         kwargs:
           gamma: [1.0, 1.0, 1.0]
