+++ satpy/etc/enhancements/amsr2.yaml
@@ -6,28 +6,28 @@ enhancements:
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 0.5}
   gaasp_sst:
     name: SST
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: -5.0, max_stretch: 35}
   gaasp_tpw:
     name: TPW
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 75.0}
   gaasp_wspd:
     name: WSPD
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 100.0}
   # Snow_Cover unscaled (category product)
   gaasp_snow_depth:
@@ -35,54 +35,54 @@ enhancements:
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 150.0}
   gaasp_swe:
     name: SWE
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 16.0}
   gaasp_soil_moisture:
     name: Soil_Moisture
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 100.0}
   gaasp_ice_concentration_nh:
     name: NASA_Team_2_Ice_Concentration_NH
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 100.0}
   gaasp_ice_concentration_sh:
     name: NASA_Team_2_Ice_Concentration_SH
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 100.0}
 #  gaasp_latency_nh:
 #    name: Latency_NH
 #    sensor: amsr2
 #    operations:
 #      - name: linear_stretch
-#        method: !!python/name:satpy.enhancements.stretching.stretch
+#        method: !!python/name:satpy.enhancements.contrast.stretch
 #        kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 100.0}
 #  gaasp_latency_sh:
 #    name: Latency_SH
 #    sensor: amsr2
 #    operations:
 #      - name: linear_stretch
-#        method: !!python/name:satpy.enhancements.stretching.stretch
+#        method: !!python/name:satpy.enhancements.contrast.stretch
 #        kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 100.0}
   gaasp_rain_rate:
     name: Rain_Rate
     sensor: amsr2
     operations:
       - name: linear_stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs: {stretch: 'crude', min_stretch: 0.0, max_stretch: 65.}
