+++ satpy/etc/enhancements/glm.yaml
@@ -12,7 +12,7 @@ enhancements:
     standard_name: c14_flash_extent_density
     operations:
       - name: stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs:
           stretch: crude
           min_stretch: [0, 0, 0]
@@ -21,7 +21,7 @@ enhancements:
     standard_name: c14_yellow_lightning
     operations:
       - name: stretch
-        method: !!python/name:satpy.enhancements.stretching.stretch
+        method: !!python/name:satpy.enhancements.contrast.stretch
         kwargs:
           stretch: crude
           min_stretch: [0, 0, 0, 0]
