+++ satpy/etc/enhancements/mersi-2.yaml
@@ -2,19 +2,19 @@ enhancements:
   default:
     operations:
     - name: stretch
-      method: !!python/name:satpy.enhancements.stretching.stretch
+      method: !!python/name:satpy.enhancements.contrast.stretch
       kwargs: {stretch: linear}
 
   cloudtop_default:
     standard_name: cloudtop
     operations:
     - name: stretch
-      method: !!python/name:satpy.enhancements.stretching.stretch
+      method: !!python/name:satpy.enhancements.contrast.stretch
       kwargs:
         stretch: crude
         min_stretch: [320, 310, 310]
         max_stretch: [220, 220, 220]
     - name: gamma
-      method: !!python/name:satpy.enhancements.stretching.gamma
+      method: !!python/name:satpy.enhancements.contrast.gamma
       kwargs:
         gamma: [1.4, 1.4, 1.2]
