+++ satpy/etc/enhancements/mhs.yaml
@@ -4,26 +4,26 @@ enhancements:
     standard_name: mw183_humidity
     operations:
     - name: inverse
-      method: !!python/name:satpy.enhancements.stretching.invert
+      method: !!python/name:satpy.enhancements.contrast.invert
       args:
       - [true, true, true]
     - name: stretch
-      method: !!python/name:satpy.enhancements.stretching.stretch
+      method: !!python/name:satpy.enhancements.contrast.stretch
       kwargs: {stretch: linear}
     - name: gamma
-      method: !!python/name:satpy.enhancements.stretching.gamma
+      method: !!python/name:satpy.enhancements.contrast.gamma
       kwargs: {gamma: 1.2}
 
   mw183_humidity_surface:
     standard_name: mw183_humidity_surface
     operations:
     - name: inverse
-      method: !!python/name:satpy.enhancements.stretching.invert
+      method: !!python/name:satpy.enhancements.contrast.invert
       args:
       - [true, true, true]
     - name: stretch
-      method: !!python/name:satpy.enhancements.stretching.stretch
+      method: !!python/name:satpy.enhancements.contrast.stretch
       kwargs: {stretch: linear}
     - name: gamma
-      method: !!python/name:satpy.enhancements.stretching.gamma
+      method: !!python/name:satpy.enhancements.contrast.gamma
       kwargs: {gamma: 1.2}
