+++ satpy/writers/ninjogeotiff.py
@@ -67,7 +67,7 @@ Moreover, two keyword arguments are no longer supported because
 their functionality has become redundant.  This applies to
 ``ch_min_measurement_unit`` and ``ch_max_measurement_unit``.
 Instead, pass those values in source units to the
-:func:`~satpy.enhancements.stretching.stretch` enhancement with the ``min_stretch``
+:func:`~satpy.enhancements.contrast.stretch` enhancement with the ``min_stretch``
 and ``max_stretch`` arguments.
 
 For images where the pixel value corresponds directly to a physical value,
